Rename the template
