+++ src/Templates/NServiceBusDockerContainer/Dockerfile
@@ -3,10 +3,10 @@ FROM microsoft/dotnet:2.0-runtime AS base
 FROM microsoft/dotnet:2.0-sdk AS build
 WORKDIR /src
 COPY . .
-WORKDIR /src/NServiceBusDockerEndpoint
+WORKDIR /src/NServiceBusDockerContainer
 RUN dotnet publish -c Release -o /app
 
 FROM base AS final
 WORKDIR /app
 COPY --from=build /app .
-ENTRYPOINT ["dotnet", "NServiceBusDockerEndpoint.dll"]
+ENTRYPOINT ["dotnet", "NServiceBusDockerContainer.dll"]
