+++ src/Templates/NServiceBusDockerContainer/NServiceBusDockerContainer.csproj
