+++ src/Templates/NServiceBusDockerContainer/Program.cs
@@ -3,7 +3,7 @@ using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace NServiceBusDockerEndpoint
+namespace NServiceBusDockerContainer
 {
     class Program
     {
