+++ src/Templates/NServiceBusDockerContainer/license.xml
